requirements.txt
pycryptodome
slixmpp
